Add debug config
