+++ api/conda_env.yml
@@ -1,0 +1,8 @@
+name: bia-integrator-api
+
+channels:
+  - default
+  - conda-forge
+
+dependencies:
+  - python=3.10
