--- core/bia_integrator_core/api/admin.py
@@ -1,11 +1,0 @@
-from fastapi import APIRouter
-
-router = APIRouter(prefix="/admin")
-
-@router.get("/health-check")
-def health_check():
-    pass
-
-@router.get("/pipelines")
-def running_pipelines():
-    pass
