+++ core/bia_integrator_core/models.py
@@ -1,8 +1,10 @@
-from pydantic import BaseModel
+import pathlib
 from typing import Dict, List, Optional, Set, Union, AnyStr
+
+from pydantic import BaseModel
+from urllib.parse import urlparse, urlunparse
 from pathlib import Path
 from ome_types import OME, from_xml
-from urllib.parse import urlparse, urlunparse
 import requests
 
 class BIABaseModel(BaseModel):
@@ -11,53 +13,19 @@ class BIABaseModel(BaseModel):
 
         return super().json(ensure_ascii=ensure_ascii, **kwargs)
 
-class Author(BIABaseModel):
-    name: str
-
-class BIAStudy(BIABaseModel):
-    id: str
-    title: str
-    description: str
-    authors: Optional[List[Author]] = []
-    organism: str
-    release_date: str
-    
-    # FIXME - this should be a list
-    imaging_type: Optional[str]
-    attributes: Dict = {}
-    example_image_uri: str = ""
-    tags: Set[str] = set()
-
-    file_references_count: int = 0
-    images_count: int = 0
-
-class FileReference(BIABaseModel):
-    """A reference to an externally hosted file."""
-
-    id: str # A unique identifier for the file reference
-    study_id: str
-    name: str # A short descriptive name
-    uri: str # URI of the file
-    size_bytes: Optional[int] # Size of the file
-    attributes: Dict = {}
 
 class ChannelRendering(BIABaseModel):
     colormap_start: List[float]
     colormap_end: List[float]
     scale_factor: float = 1.0
 
+
 class RenderingInfo(BIABaseModel):
     channel_renders: List[ChannelRendering]
     default_z: Optional[int]
     default_t: Optional[int]
 
-class BIAImageAlias(BIABaseModel):
-    """An alias for an image - a more convenient way to refer to the image than
-    the full accession ID / UUID pair"""
 
-    name: str
-    accession_id: str
-    
 class BIAImageRepresentation(BIABaseModel):
     """A particular representation of a BIAImage. Examples:
     
@@ -67,7 +35,7 @@ class BIAImageRepresentation(BIABaseModel):
     * A thumbnail."""
     
     accession_id: str
-    image_aliases: List[BIAImageAlias]
+    image_id: str
     uri: Union[str, List[str]]
     size: int
     type: Optional[str]
@@ -75,22 +43,31 @@ class BIAImageRepresentation(BIABaseModel):
     attributes: Optional[Dict]
     rendering: Optional[RenderingInfo]
 
+
 class BIAFileRepresentation(BIABaseModel):
     accession_id: str
     file_id: str
     uri: Union[str, List[str]]
     size: int
 
+
 class BIAFile(BIABaseModel):
     id: str
-    study_id: str
-    original_relpath: Path
+    original_relpath: pathlib.Path
     original_size: int
     attributes: Dict = {}
     representations: List[BIAFileRepresentation] = []
 
-class BIAOmeMetadata(dict):
-    pass
+
+class FileReference(BIABaseModel):
+    """A reference to an externally hosted file."""
+
+    id: str # A unique identifier for the file reference
+    name: str # A short descriptive name
+    uri: str # URI of the file
+    size_in_bytes: Optional[int] # Size of the file
+    attributes: Dict = {}
+
 
 class BIAImage(BIABaseModel):
     """This class represents the abstract concept of an image. Images are
@@ -108,18 +85,18 @@ class BIAImage(BIABaseModel):
     """
 
     id: str
-    study_id: str
-    accession_id: str
+    # TODO - this should be mandatory, but it's a breaking change
+    accession_id: Optional[str] = None
     # TODO - rationalise these (name should probably replace original_relpath)
     name: Optional[str]
-    original_relpath: Path
+    original_relpath: pathlib.Path
 
     dimensions: Optional[str]
     representations: List[BIAImageRepresentation] = []
     attributes: Dict = {}
 
     @property
-    def ome_metadata(self) -> Optional[BIAOmeMetadata]:
+    def ome_metadata(self) -> Optional[OME]:
         metadata = self.__dict__.get('ome_metadata', None)
         if metadata is None:
             ngff_rep = [rep for rep in self.representations if rep.type == "ome_ngff"]
@@ -151,6 +128,43 @@ class BIAImage(BIABaseModel):
 
         return ome_metadata
 
+class BIAImageAlias(BIABaseModel):
+    """An alias for an image - a more convenient way to refer to the image than
+    the full accession ID / UUID pair"""
+
+    name: str
+    accession_id: str
+    image_id: str
+
+
+class Author(BIABaseModel):
+    name: str
+
+
+class BIAStudy(BIABaseModel):
+    accession_id: str
+    title: str
+    description: str
+    authors: Optional[List[Author]] = []
+    organism: str
+    release_date: str
+    
+    # FIXME - this should be a list
+    imaging_type: Optional[str]
+    attributes: Dict = {}
+    example_image_uri: str = ""
+
+    file_references: Dict[str, FileReference] = {}
+
+    images: Dict[str, BIAImage] = {}
+    archive_files: Dict[str, BIAFile] = {}
+    other_files: Dict[str, BIAFile] = {}
+
+    image_aliases: Dict[str, BIAImageAlias] = {}
+
+    tags: Set[str] = set()
+
+
 class BIACollection(BIABaseModel):
     """A collection of studies with a coherent purpose. Studies can be in 
     multiple collections."""
@@ -161,14 +175,17 @@ class BIACollection(BIABaseModel):
     description: Optional[str]
     accession_ids: List[str]
 
+
 class StudyTag(BIABaseModel):
     accession_id: str
     value: str
+
 
 class StudyAnnotation(BIABaseModel):
     accession_id: str
     key: str
     value: str
+
 
 class ImageAnnotation(BIABaseModel):
     accession_id: str
