--- core/bia_integrator_core/repository.py
@@ -1,16 +1,0 @@
-import bia_integrator_core.models as models
-
-def file_references_for_study(study_id: str):
-    pass
-
-def images_for_study(study_id: str):
-    pass
-
-def find_image_by_id(image_id: str):
-    pass
-
-def refresh_counts(image_id: str):
-    pass
-
-def persist_study(study: models.BIAStudy) -> models.BIAStudy:
-    pass
